+++ CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog][kac] and this project adheres to
 [kac]: http://keepachangelog.com/en/1.0.0/
 [semver]: http://semver.org/spec/v2.0.0.html
 
+# _NEXT_
+
+## Added
+
+* `ConfigFile` validates before writing out the configuration.
+* `ConfigFile` exposes the `fileName` property.
+
 # v0.4.0 (2018-04-11)
 
 ## Added